Create a more sensible interface (IMO) to the SimpleNote API
It's a little bit of wheel re-invention but I think I'm starting to get
what I'm doing with Android and the way I want to approach things is
diverging from the way Josh and Ben approached things quite a bit. This
is the start of my code branching off the way my thought seems to have.