Set a boolean when saving state and when resuming
The boolean identifies whether we left due to user action like the back
button or due to an action that expects a return to edit when simplenote
gets to the top of the Activity stack again.
Move methods so they are in the order they get fired in the Activity
lifecycle.