Make blue background look pressed when touching trash button
Add the resources to change the way the note title is displayed in the
note editing activity. Note title is just the arrow with transparency
that sits on top of the LinearLayout's blue background. The trash icon
has an empty background.
Add an onTouch event listener to the trash icon that sets the
LinearLayout to pressed when the trash icon is touched.
[#63 tagged:ui state:resolved]