Add dropshadow under note list if notes don't fill list view
If the note list fills the height of the display (minus padding and
scroll shadow) then set the list padding to zero (which hides the drop
shadow) otherwise set padding to 29dip. The 29 comes from an integer
value in the integers.xml. The list height calcuation is based on an
integer value in integers.xml as well because figuring out when one is
drawn to measure it is hard.
[#65 state:resolved]